+++ NEWS.md
@@ -2,3 +2,9 @@
 # News
 
 ## v0.8.0
+
+- Implement `show(::IO, ::TypstCommand)`
+
+### Bug fixes
+
+- Fixed `show(::IO, ::MIME"text/plain", ::TypstCommand)` failing to print to the `IO`